requirements.txt
# Reddit Market Radar v2.0 - Bağımlılıklar

# Temel
requests>=2.28.0
python-dotenv>=1.0.0

# AI Sağlayıcılar (birini veya ikisini de kullanabilirsiniz)
openai>=1.0.0
google-generativeai>=0.3.0

# PRAW - Reddit API (isteğe bağlı, şu an HTTP ile çalışıyor)
praw>=7.7.0

# Yardımcı
typing_extensions>=4.0.0
